Fix issue #11 Test Id can also be used to match test cases
When updating the test run status the lookup was based on matching the test case name instead of the JAMA ID (DocumentKey). This improvement allows users to identify the test cases in Allure (TmsLink) with the name OR the JAMA ID, allowing a much better matching. It has the additional cost of one more call to the JAMA API per test case.